requirements.txt
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
tensorflow==2.16.1
opencv-python==4.9.0.80
Pillow==10.2.0
scikit-learn==1.4.1.post1
seaborn==0.13.2
